Renamed pafi.py for clarity
