+++ pafi/__init__.py
@@ -1,1 +1,1 @@
-from pafi.pafi import *
+from pafi.post_processing import *
